support choice to delete, or not, merged customer. moved merge button

+++ src/Libraries/Nop.Services/Customers/CustomerService.cs
@@ -307,15 +307,15 @@ public partial class CustomerService : ICustomerService
         //filter customers by billing country
         if (countryId > 0)
             customers = from c in customers
-                join a in _customerAddressRepository.Table on c.BillingAddressId equals a.Id
-                where a.CountryId == countryId
-                select c;
+                        join a in _customerAddressRepository.Table on c.BillingAddressId equals a.Id
+                        where a.CountryId == countryId
+                        select c;
 
         var customersWithCarts = from c in customers
-            join item in items on c.Id equals item.CustomerId
-            //we change ordering for the MySQL engine to avoid problems with the ONLY_FULL_GROUP_BY server property that is set by default since the 5.7.5 version
-            orderby _dataProvider.ConfigurationName == "MySql" ? c.CreatedOnUtc : item.CreatedOnUtc descending
-            select c;
+                                 join item in items on c.Id equals item.CustomerId
+                                 //we change ordering for the MySQL engine to avoid problems with the ONLY_FULL_GROUP_BY server property that is set by default since the 5.7.5 version
+                                 orderby _dataProvider.ConfigurationName == "MySql" ? c.CreatedOnUtc : item.CreatedOnUtc descending
+                                 select c;
 
         return await customersWithCarts.Distinct().ToPagedListAsync(pageIndex, pageSize);
     }
@@ -401,8 +401,8 @@ public partial class CustomerService : ICustomerService
             return null;
 
         var query = from c in _customerRepository.Table
-            where customerGuids.Contains(c.CustomerGuid)
-            select c;
+                    where customerGuids.Contains(c.CustomerGuid)
+                    select c;
         var customers = await query.ToListAsync();
 
         return customers;
@@ -422,9 +422,9 @@ public partial class CustomerService : ICustomerService
             return null;
 
         var query = from c in _customerRepository.Table
-            where c.CustomerGuid == customerGuid
-            orderby c.Id
-            select c;
+                    where c.CustomerGuid == customerGuid
+                    orderby c.Id
+                    select c;
 
         return await _shortTermCacheManager.GetAsync(async () => await query.FirstOrDefaultAsync(), NopCustomerServicesDefaults.CustomerByGuidCacheKey, customerGuid);
     }
@@ -443,9 +443,9 @@ public partial class CustomerService : ICustomerService
             return null;
 
         var query = from c in _customerRepository.Table
-            orderby c.Id
-            where c.Email == email
-            select c;
+                    orderby c.Id
+                    where c.Email == email
+                    select c;
         var customer = await query.FirstOrDefaultAsync();
 
         return customer;
@@ -465,9 +465,9 @@ public partial class CustomerService : ICustomerService
             return null;
 
         var query = from c in _customerRepository.Table
-            orderby c.Id
-            where c.SystemName == systemName
-            select c;
+                    orderby c.Id
+                    where c.SystemName == systemName
+                    select c;
 
         var customer = await _shortTermCacheManager.GetAsync(async () => await query.FirstOrDefaultAsync(), NopCustomerServicesDefaults.CustomerBySystemNameCacheKey, systemName);
 
@@ -564,9 +564,9 @@ public partial class CustomerService : ICustomerService
             return null;
 
         var query = from c in _customerRepository.Table
-            orderby c.Id
-            where c.Username == username
-            select c;
+                    orderby c.Id
+                    where c.Username == username
+                    select c;
         var customer = await query.FirstOrDefaultAsync();
 
         return customer;
@@ -682,28 +682,28 @@ public partial class CustomerService : ICustomerService
         var guestRole = await GetCustomerRoleBySystemNameAsync(NopCustomerDefaults.GuestsRoleName);
 
         var allGuestCustomers = from guest in _customerRepository.Table
-            join ccm in _customerCustomerRoleMappingRepository.Table on guest.Id equals ccm.CustomerId
-            where ccm.CustomerRoleId == guestRole.Id
-            select guest;
+                                join ccm in _customerCustomerRoleMappingRepository.Table on guest.Id equals ccm.CustomerId
+                                where ccm.CustomerRoleId == guestRole.Id
+                                select guest;
 
         var guestsToDelete = from guest in _customerRepository.Table
-            join g in allGuestCustomers on guest.Id equals g.Id
-            from sCart in _shoppingCartRepository.Table.Where(sci => sci.CustomerId == guest.Id).DefaultIfEmpty()
-            from order in _orderRepository.Table.Where(o => o.CustomerId == guest.Id).DefaultIfEmpty()
-            from blogComment in _blogCommentRepository.Table.Where(o => o.CustomerId == guest.Id).DefaultIfEmpty()
-            from newsComment in _newsCommentRepository.Table.Where(o => o.CustomerId == guest.Id).DefaultIfEmpty()
-            from productReview in _productReviewRepository.Table.Where(o => o.CustomerId == guest.Id).DefaultIfEmpty()
-            from productReviewHelpfulness in _productReviewHelpfulnessRepository.Table.Where(o => o.CustomerId == guest.Id).DefaultIfEmpty()
-            from pollVotingRecord in _pollVotingRecordRepository.Table.Where(o => o.CustomerId == guest.Id).DefaultIfEmpty()
-            from forumTopic in _forumTopicRepository.Table.Where(o => o.CustomerId == guest.Id).DefaultIfEmpty()
-            from forumPost in _forumPostRepository.Table.Where(o => o.CustomerId == guest.Id).DefaultIfEmpty()
-            where (!onlyWithoutShoppingCart || sCart == null) &&
-                  order == null && blogComment == null && newsComment == null && productReview == null && productReviewHelpfulness == null &&
-                  pollVotingRecord == null && forumTopic == null && forumPost == null &&
-                  !guest.IsSystemAccount &&
-                  (createdFromUtc == null || guest.CreatedOnUtc > createdFromUtc) &&
-                  (createdToUtc == null || guest.CreatedOnUtc < createdToUtc)
-            select new { CustomerId = guest.Id };
+                             join g in allGuestCustomers on guest.Id equals g.Id
+                             from sCart in _shoppingCartRepository.Table.Where(sci => sci.CustomerId == guest.Id).DefaultIfEmpty()
+                             from order in _orderRepository.Table.Where(o => o.CustomerId == guest.Id).DefaultIfEmpty()
+                             from blogComment in _blogCommentRepository.Table.Where(o => o.CustomerId == guest.Id).DefaultIfEmpty()
+                             from newsComment in _newsCommentRepository.Table.Where(o => o.CustomerId == guest.Id).DefaultIfEmpty()
+                             from productReview in _productReviewRepository.Table.Where(o => o.CustomerId == guest.Id).DefaultIfEmpty()
+                             from productReviewHelpfulness in _productReviewHelpfulnessRepository.Table.Where(o => o.CustomerId == guest.Id).DefaultIfEmpty()
+                             from pollVotingRecord in _pollVotingRecordRepository.Table.Where(o => o.CustomerId == guest.Id).DefaultIfEmpty()
+                             from forumTopic in _forumTopicRepository.Table.Where(o => o.CustomerId == guest.Id).DefaultIfEmpty()
+                             from forumPost in _forumPostRepository.Table.Where(o => o.CustomerId == guest.Id).DefaultIfEmpty()
+                             where (!onlyWithoutShoppingCart || sCart == null) &&
+                                   order == null && blogComment == null && newsComment == null && productReview == null && productReviewHelpfulness == null &&
+                                   pollVotingRecord == null && forumTopic == null && forumPost == null &&
+                                   !guest.IsSystemAccount &&
+                                   (createdFromUtc == null || guest.CreatedOnUtc > createdFromUtc) &&
+                                   (createdToUtc == null || guest.CreatedOnUtc < createdToUtc)
+                             select new { CustomerId = guest.Id };
 
         await using var tmpGuests = await _dataProvider.CreateTempDataStorageAsync("tmp_guestsToDelete", guestsToDelete);
         await using var tmpAddresses = await _dataProvider.CreateTempDataStorageAsync("tmp_guestsAddressesToDelete",
@@ -1215,9 +1215,9 @@ public partial class CustomerService : ICustomerService
         var key = _staticCacheManager.PrepareKeyForDefaultCache(NopCustomerServicesDefaults.CustomerRolesBySystemNameCacheKey, systemName);
 
         var query = from cr in _customerRoleRepository.Table
-            orderby cr.Id
-            where cr.SystemName == systemName
-            select cr;
+                    orderby cr.Id
+                    where cr.SystemName == systemName
+                    select cr;
 
         var customerRole = await _staticCacheManager.GetAsync(key, async () => await query.FirstOrDefaultAsync());
 
@@ -1615,9 +1615,9 @@ public partial class CustomerService : ICustomerService
     public virtual async Task<IList<Address>> GetAddressesByCustomerIdAsync(int customerId)
     {
         var query = from address in _customerAddressRepository.Table
-            join cam in _customerAddressMappingRepository.Table on address.Id equals cam.AddressId
-            where cam.CustomerId == customerId
-            select address;
+                    join cam in _customerAddressMappingRepository.Table on address.Id equals cam.AddressId
+                    where cam.CustomerId == customerId
+                    select address;
 
         return await _shortTermCacheManager.GetAsync(async () => await query.ToListAsync(), NopCustomerServicesDefaults.CustomerAddressesCacheKey, customerId);
     }
@@ -1637,9 +1637,9 @@ public partial class CustomerService : ICustomerService
             return null;
 
         var query = from address in _customerAddressRepository.Table
-            join cam in _customerAddressMappingRepository.Table on address.Id equals cam.AddressId
-            where cam.CustomerId == customerId && address.Id == addressId
-            select address;
+                    join cam in _customerAddressMappingRepository.Table on address.Id equals cam.AddressId
+                    where cam.CustomerId == customerId && address.Id == addressId
+                    select address;
 
         return await _shortTermCacheManager.GetAsync(async () => await query.FirstOrDefaultAsync(), NopCustomerServicesDefaults.CustomerAddressCacheKey, customerId, addressId);
     }
@@ -1723,7 +1723,8 @@ public partial class CustomerService : ICustomerService
         if (anyRolesModified)
             await _staticCacheManager.RemoveByPrefixAsync(NopCustomerServicesDefaults.CustomerCustomerRolesPrefix);
 
-        await _customerRepository.DeleteAsync(fromCustomer);
+        if (deleteFromCustomer)
+            await _customerRepository.DeleteAsync(fromCustomer);
 
     }
